@@ -3,16 +3,19 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint"
+    "dev": "vite",
+    "build": "vue-tsc && vite build",
+    "preview": "vite preview"
   },
   "dependencies": {
+    "@vitejs/plugin-vue": "^4.2.3",
+    "@vitejs/plugin-vue-jsx": "^3.0.1",
     "axios": "^1.4.0",
     "core-js": "^3.8.3",
     "pinia": "^2.1.4",
     "sass": "^1.62.1",
     "sass-loader": "^13.3.0",
+    "vite": "^4.4.9",
     "vue": "^3.2.13",
     "vue-router": "^4.2.1",
     "vue-toastification": "^2.0.0-rc.5"
@@ -26,9 +29,10 @@
     "@vue/cli-plugin-eslint": "~5.0.0",
     "@vue/cli-plugin-typescript": "~5.0.0",
     "@vue/cli-service": "~5.0.0",
-    "@vue/eslint-config-typescript": "^9.1.0",
-    "eslint": "^7.32.0",
-    "eslint-plugin-vue": "^8.0.3",
+    "@vue/eslint-config-typescript": "^11.0.3",
+    "eslint": "^8.46.0",
+    "eslint-plugin-vue": "^9.16.0",
+    "eslint-plugin-vue-scoped-css": "^2.5.0",
     "typescript": "~4.5.5"
   },
   "eslintConfig": {
