@@ -1,16 +1,17 @@
 <template>
+  <Header />
   <router-view></router-view>
 </template>
 
-<script lang="ts" setup></script>
+<script lang="ts" setup>
+import Header from "@/components/Header.vue";
+</script>
 
 <style lang="scss">
+body {
+  margin: 0;
+  padding: 0;
+}
 #app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-  text-align: center;
-  color: #2c3e50;
-  margin-top: 60px;
 }
 </style>
