@@ -1,7 +1,7 @@
 import { createRouter, createWebHistory } from "vue-router";
 import RegistrationPage from "../components/RegistrationPage.vue";
-import LoginPage from "../components/LoginPage.vue";
-import MainPage from "../components/MainPage.vue";
+import LoginPage from "@/components/LoginPage.vue";
+import MainPage from "@/components/RegistrationPage.vue";
 
 const routes = [
   { path: "/registration", component: RegistrationPage },
